move __connect_awaitable into its own header
