+++ include/stdexec/__detail/__awaitable.hpp
@@ -18,19 +18,8 @@
 #include "__config.hpp"
 #include "__concepts.hpp"
 
-#include <version>
-#if __cpp_impl_coroutine >= 201902 && __cpp_lib_coroutine >= 201902
-#  include <coroutine>
-namespace __coro = std;
-#elif defined(__cpp_coroutines) && __has_include(<experimental/coroutine>)
-#  include <experimental/coroutine>
-namespace __coro = std::experimental;
-#else
-#  define STDEXEC_STD_NO_COROUTINES_ 1
-#endif
-
 namespace stdexec {
-#if !STDEXEC_STD_NO_COROUTINES_
+#if !STDEXEC_STD_NO_COROUTINES()
   // Define some concepts and utilities for working with awaitables
   template <class _Tp>
   concept __await_suspend_result =
